patch_extraction/requirements.txt
bokeh==2.4.2
certifi==2022.6.15
cloudpickle==2.0.0
colorcet==3.0.0
cycler==0.11.0
dask==2022.1.0
datashader==0.13.0
datashape==0.5.2
distributed==2022.1.0
fonttools==4.29.0
fsspec==2022.1.0
HeapDict==1.0.1
histolab==0.5.1
holoviews==1.14.7
imageio==2.14.1
kiwisolver==1.3.2
llvmlite==0.38.0
locket==0.2.1
matplotlib==3.5.1
msgpack==1.0.3
multipledispatch==0.6.0
networkx==2.8.5
numba==0.55.0
numpy==1.20.0
openslide-python==1.2.0
packaging==21.3
panel==0.12.6
param==1.12.0
partd==1.2.0
Pillow==8.4.0
psutil==5.9.0
pyct==0.4.8
pyparsing==3.0.9
pyviz-comms==2.1.0
PyWavelets==1.2.0
PyYAML==6.0
scikit-image==0.19.1
scipy==1.8.0
seaborn==0.11.2
sortedcontainers==2.4.0
tblib==1.7.0
tifffile==2021.11.2
toolz==0.11.2
torch==1.4.0
torchvision==0.5.0
typing_extensions==4.3.0
xarray==0.20.2
zict==2.0.0
